feat(ui): move List to experimental and drop its stories for now
The new List module is the first entry under ui/src/experimental,
exported as ./experimental/List: its API is not settled until the desk
list page has consumed it, and on develop the package is visible to
every app. The three .story.vue files go with it; how this package shows
a component's stories is an open question, and they return with its
answer.
