+++ ui/CONTEXT.md
@@ -77,8 +77,10 @@ together against a real doctype to verify pixel parity with CRM's list view.
 Distinct from each control's own isolated story.
 
 **List module**:
-The shared module that holds the table (`List`), the footer (`ListFooter`) and the
-selection bar (`ListBulkBar`) a doctype list page composes beside the controls. The
+The experimental module (`experimental/List/`, subpath `./experimental/List`) that holds
+the table (`List`), the footer (`ListFooter`) and the selection bar (`ListBulkBar`) a
+doctype list page composes beside the controls. Experimental: its API can change with
+no deprecation window until the desk list page has settled it. The
 host lays them out, fetches the rows, and binds the same **Column**s and **Sort**s the
 controls edit; the table itself fetches and stores nothing.
