+++ ui/docs/adr/0002-listview-module-layout.md
@@ -1,7 +1,7 @@
 # List View module layout: one module per control + a composite module
 
-> Note (2026-09-09): the table, footer and bulk bar landed as a separate `List` module (see
-> `CONTEXT.md`, "List module"), not inside the composite. The composite `ListView/` module is
+> Note (2026-09-09): the table, footer and bulk bar landed as a separate experimental `List` module
+> (`src/experimental/List/`, see `CONTEXT.md`, "List module"), not inside the composite. The composite `ListView/` module is
 > kept for now and is not the desk list page's host.
 
 Each extracted control gets its own `@framework/ui` module folder (`SortBy/`,
