+++ ui/package.json
@@ -25,9 +25,9 @@
       "import": "./src/components/InviteUser/index.ts",
       "types": "./src/components/InviteUser/index.ts"
     },
-    "./List": {
-      "import": "./src/components/List/index.ts",
-      "types": "./src/components/List/index.ts"
+    "./experimental/List": {
+      "import": "./src/experimental/List/index.ts",
+      "types": "./src/experimental/List/index.ts"
     },
     "./ListView": {
       "import": "./src/components/ListView/index.ts",
